Add click handler

+++ src/main/java/scc210game/state/event/MouseMovedEvent.java
@@ -1,10 +1,10 @@
 package scc210game.state.event;
 
 public class MouseMovedEvent implements InputEvent {
-    public final int dx;
-    public final int dy;
+    public final float dx;
+    public final float dy;
 
-    public MouseMovedEvent(int dx, int dy) {
+    public MouseMovedEvent(float dx, float dy) {
         this.dx = dx;
         this.dy = dy;
     }
